consider moving compute_scal and into a preprocessing module
In a refactored version of this codebase, should there be a preprocessing
module with functions for ETL and formatting data for the transport map?
Some examples would be tools to aid reading netcdfc files, parsing dates
(for subsetting large netcdf formats), and reshaping datasets before loading
them in the transport map.