First experiments with fitting covariates (all zero)
Linear fits do not work and throw an error when estimating a Cholesky
factor.  I have not investigated this much yet and do not have any
hypotheses about why the error occurs there.

Nonlinear fits work but throw an error when drawing conditional
samples.  The error traces from a cStar call to the kernel_fun,
where the issue (running example1 adapted with covariates) is that
we try torch.cat((Y, X), 1) to concatenate Y and X columnwise.  At
the error Y has dim (1 x 15) and X dim (10 x 1). The difference
seems to be that Y is using only one ensemble member at a time, or
that Y is essentially averaged over ensmembers before conditional
sampling.  X on the other hand is still using the input data X from
the start of the script.  We know that the fit method and kernel_fun
work because the error occurs when calling cond_sample.