enables importing package from other dirs, simplifies env setup
This commit provides an __init__ file so we can import the package
in other directories. The Python sys library allows to provide a
path to this repo locally for import.

It also furnishes a yaml file that can simplify environment setup
using miniconda3 or the anaconda package manager (the same thing).

Contact me <danjdrennan> for assistance with environtment setup or
importing parts of this package from other directories. This change
makes working in a Jupyter notebook, for example, a lot easier.